[SYCL] Move AreAllButLastReductions to reduction_forward.hpp
Otherwise, `<sycl/detail/core.hpp>` isn't enough to invoke non-reduction
`parallel_for`.
