+++ sycl/include/sycl/reduction.hpp
@@ -60,32 +60,6 @@
 
 namespace sycl {
 inline namespace _V1 {
-namespace detail {
-
-/// Base non-template class which is a base class for all reduction
-/// implementation classes. It is needed to detect the reduction classes.
-class reduction_impl_base {};
-
-/// Predicate returning true if a type is a reduction.
-template <typename T> struct IsReduction {
-  static constexpr bool value =
-      std::is_base_of_v<reduction_impl_base, std::remove_reference_t<T>>;
-};
-
-/// Predicate returning true if all template type parameters except the last one
-/// are reductions.
-template <typename FirstT, typename... RestT> struct AreAllButLastReductions {
-  static constexpr bool value =
-      IsReduction<FirstT>::value && AreAllButLastReductions<RestT...>::value;
-};
-
-/// Helper specialization of AreAllButLastReductions for one element only.
-/// Returns true if the template parameter is not a reduction.
-template <typename T> struct AreAllButLastReductions<T> {
-  static constexpr bool value = !IsReduction<T>::value;
-};
-} // namespace detail
-
 /// Class that is used to represent objects that are passed to user's lambda
 /// functions and representing users' reduction variable.
 /// The generic version of the class represents those reductions of those
