+++ sycl/test-e2e/AddressCast/static_address_cast.cpp
@@ -10,7 +10,10 @@
 // UNSUPPORTED: hip, cuda
 // RUN: %{build} -o %t.out
 // RUN: %{run} %t.out
-#include <sycl/sycl.hpp>
+
+#include <sycl/detail/core.hpp>
+
+#include <sycl/ext/oneapi/experimental/address_cast.hpp>
 
 int main() {
